requirements/base.txt
# This file collects all required third-party applications that are needed
# to run this project. Later you can install all these apps in a row
# using pip. Example::
#
#     pip install -r requirements.pip

Django==1.6
South<=2.0
MySQL-python==1.2.4
Pillow==2.1.0
django-autoslug==1.7.1
django-bootstrap3==2.0.0
django-braces==1.2.2
django-crispy-forms==1.4.0
django-grappelli==2.5.1
django-taggit==0.10
python-memcached==1.52
unicode-slugify==0.1.1
